% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_viruses_by_group}
\alias{get_viruses_by_group}
\title{Get Viruses by Sample Group}
\usage{
get_viruses_by_group(tpm_mat, sample_groups, group_name)
}
\arguments{
\item{tpm_mat}{A numeric matrix of TPM values (rows = vOTUs, columns = samples).}

\item{sample_groups}{A named character vector mapping sample IDs to group names.}

\item{group_name}{Character. The target group name to filter.}
}
\value{
Character vector of vOTU IDs present in the specified group.
}
\description{
Returns virus IDs present (TPM > 0) in a specified sample group.
}
\examples{
\dontrun{
bs_viruses <- get_viruses_by_group(tpm_mat, sample_groups, "BS")
}

}
