% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_fig2}
\alias{theme_fig2}
\title{Custom ggplot2 Theme for Figure 2}
\usage{
theme_fig2()
}
\value{
A ggplot2 theme object.
}
\description{
A consistent minimal theme for Figure 2 panels with Times font family.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_fig2()
}

}
